+++ src/Components/index.js
@@ -1,0 +1,23 @@
+import CartCard from './CartCard';
+import CartPriceCard from './CartPriceCard';
+import Catog from './Catog';
+import FilterBar from './FilterBar';
+import Footer from './Footer';
+import HeroSlider from './HeroSlider';
+import Navbar from './Navbar';
+import ProductCard from './ProductCard';
+import Offers from './Offers';
+import EmptyCart from './CartComponents/EmptyCart';
+
+export {
+	CartCard,
+	Catog,
+	FilterBar,
+	Footer,
+	HeroSlider,
+	Navbar,
+	ProductCard,
+	CartPriceCard,
+	Offers,
+	EmptyCart,
+};
