+++ src/pages/HomePage.jsx
@@ -1,12 +1,13 @@
-import Navbar from "../Components/Navbar"
-import HeroSlider from "../Components/HeroSlider"
-import Footer from "../Components/Footer"
-import Offers from "../Components/Offers"
-import Catog from "../Components/Catog"
-
+import { Navbar, HeroSlider, Footer, Offers, Catog } from '../Components/index';
 const HomePage = () => {
-    return (
-        <div><Navbar /><HeroSlider /><Offers /><Catog /><Footer /></div>
-    ) 
-}
+	return (
+		<div>
+			<Navbar />
+			<HeroSlider />
+			<Offers />
+			<Catog />
+			<Footer />
+		</div>
+	);
+};
 export default HomePage;
