+++ src/pages/Login.jsx
@@ -1,7 +1,6 @@
-import React from 'react';
 import { Link } from 'react-router-dom';
-import Footer from '../Components/Footer';
-import Navbar from '../Components/Navbar';
+import { Footer, Navbar } from '../Components/index';
+
 function Login() {
 	return (
 		<div>
